Refactor NFS-related functionality to a separate file
